+++ specification/_global/get_script/GetScriptRequest.ts
@@ -27,11 +27,11 @@ import { Time } from '@_types/Time'
  * @stability TODO
  */
 export interface Request extends RequestBase {
-  path_parts?: {
+  path_parts: {
     id: Id
   }
   query_parameters?: {
+    /** Specify timeout for connection to master */
     master_timeout?: Time
   }
-  body?: {}
 }
